fix: run crossorigin injection after jampack via standalone script
Agent-Logs-Url: https://github.com/Atlas-OS/website/sessions/a9f6c263-7f34-4c39-b01b-d651c419828d
